Update naming to match path convention
